pyproject.toml
[tool.poetry]
name = "ninja"
version = "0.1.0"
description = ""
authors = ["hi1d <aopd408@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
black = "21.12b0"
django-stubs = "1.9.0"
mypy = "0.931"
isort = "5.10.1"
django-ninja = "0.16.1"
PyMySQL = "1.0.2"
types-PyMySQL = "1.0.6"
types-setuptools = "^57.4.9"
django-extensions = "^3.1.5"
ipython = "^8.0.1"
gunicorn = "20.1.0"
uvicorn = "0.16.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


# Black
[tool.black]
line-length = 120

# Mypy
[tool.mypy]
plugins = ["mypy_django_plugin.main"]
python_version = 3.9
strict = true

[[tool.mypy.overrides]]
module = "*.migrations.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = "manage"
ignore_errors = true

[tool.django-stubs]
django_settings_module = "django_ninja.settings"

# isort
[tool.isort]
profile = "black"